fix(ssh): filter terminal prompt input continuously
