+++ CLI/SSHPTYAttachReconnectInputFilter.swift
@@ -1,23 +1,13 @@
+import CmuxFoundation
 import Darwin
 import Foundation
 
 final class SSHPTYAttachReconnectInputFilter {
-    private static let escape: UInt8 = 0x1B
-    private static let endOfTransmission: UInt8 = 0x04
-    private static let bell: UInt8 = 0x07
-    private static let leftBracket: UInt8 = 0x5B
-    private static let rightBracket: UInt8 = 0x5D
-    private static let backslash: UInt8 = 0x5C
-    private static let semicolon: UInt8 = 0x3B
-    private static let questionMark: UInt8 = 0x3F
-    private static let dollar: UInt8 = 0x24
-    private static let maxPendingProbeBytes = 512
     // Terminal ESC disambiguation: bounded so a literal Escape key is not held indefinitely.
     private static let pendingProbeContinuationTimeoutMilliseconds: Int32 = 25
     private static let reconnectProbeDeadlineMilliseconds: Int64 = 2_000
 
-    private var isFiltering: Bool
-    private var pending = [UInt8]()
+    private var byteFilter: SSHPTYReconnectInputByteFilter
     private let deadlineReached: (@Sendable () -> Bool)?
     private let remainingDeadline: (@Sendable () -> Int64?)?
 
@@ -26,14 +16,13 @@ final class SSHPTYAttachReconnectInputFilter {
         deadlineReached: (@Sendable () -> Bool)? = nil,
         remainingDeadlineMilliseconds: (@Sendable () -> Int64?)? = nil
     ) {
-        isFiltering = enabled
+        byteFilter = SSHPTYReconnectInputByteFilter(enabled: enabled)
         self.deadlineReached = deadlineReached
         remainingDeadline = remainingDeadlineMilliseconds
     }
 
     private init(state: SSHPTYAttachReconnectInputFilterState) {
-        isFiltering = state.isFiltering
-        pending = state.pending
+        byteFilter = SSHPTYReconnectInputByteFilter(enabled: state.isFiltering)
         deadlineReached = state.deadlineReached
         remainingDeadline = state.remainingDeadlineMilliseconds
     }
@@ -53,7 +42,6 @@ final class SSHPTYAttachReconnectInputFilter {
         let filterState = filterEnabled
             ? SSHPTYAttachReconnectInputFilterState(
                 isFiltering: true,
-                pending: [],
                 deadlineReached: { deadline.map { now() >= $0 } ?? false },
                 remainingDeadlineMilliseconds: {
                     guard let deadline else { return nil }
@@ -284,223 +272,44 @@ final class SSHPTYAttachReconnectInputFilter {
     }
 
     func filter(_ data: Data) -> Data {
-        guard isFiltering, !data.isEmpty else {
-            return data
-        }
         if isDeadlineReached {
             var output = stopFiltering()
             output.append(data)
             return output
         }
-
-        var bytes = pending
-        pending.removeAll(keepingCapacity: true)
-        bytes.append(contentsOf: data)
-
-        var output = Data()
-        var index = 0
-        while index < bytes.count {
-            if bytes[index] == Self.endOfTransmission {
-                index += 1
-                continue
-            }
-            guard bytes[index] == Self.escape else {
-                isFiltering = false
-                output.append(contentsOf: bytes[index...])
-                return output
-            }
-
-            switch Self.reconnectProbeReplySequence(in: bytes, at: index) {
-            case .strip(let length):
-                index += length
-            case .incomplete:
-                let suffix = bytes[index...]
-                guard suffix.count <= Self.maxPendingProbeBytes else {
-                    isFiltering = false
-                    output.append(contentsOf: suffix)
-                    return output
-                }
-                pending.append(contentsOf: suffix)
-                return output
-            case .passThrough:
-                isFiltering = false
-                output.append(contentsOf: bytes[index...])
-                return output
-            }
-        }
-
-        return output
+        return byteFilter.filter(data)
     }
 
     func finish() -> Data {
-        guard !pending.isEmpty else {
-            return Data()
-        }
-        let data = Data(pending)
-        pending.removeAll(keepingCapacity: false)
-        return data
+        byteFilter.finish()
     }
 
     func stopFiltering() -> Data {
-        let input = finish()
-        isFiltering = false
-        return input
+        byteFilter.stopFiltering()
     }
 
     var hasPendingInput: Bool {
-        isFiltering && !pending.isEmpty
+        byteFilter.hasPendingInput
     }
 
     var isFilteringAtProbeBoundary: Bool {
-        isFiltering && pending.isEmpty
+        byteFilter.isFilteringAtProbeBoundary
     }
 
     var isFilteringActive: Bool {
-        isFiltering
+        byteFilter.isFilteringActive
     }
 
     var isDeadlineReached: Bool {
-        isFiltering && (deadlineReached?() == true)
+        byteFilter.isFilteringActive && (deadlineReached?() == true)
     }
 
     var remainingDeadlineMilliseconds: Int64? {
-        guard isFiltering else { return nil }
+        guard byteFilter.isFilteringActive else { return nil }
         return remainingDeadline?()
     }
 
     func flushPendingInput() -> Data {
-        guard hasPendingInput else {
-            return Data()
-        }
-        let data = Data(pending)
-        pending.removeAll(keepingCapacity: true)
-        isFiltering = false
-        return data
+        byteFilter.flushPendingInput()
     }
-
-    private static func reconnectProbeReplySequence(
-        in bytes: [UInt8],
-        at start: Int
-    ) -> SSHPTYAttachReconnectInputFilterSequenceMatch {
-        guard start < bytes.count, bytes[start] == escape else {
-            return .passThrough
-        }
-        guard start + 1 < bytes.count else {
-            // read() can split immediately after ESC; wait for one more byte before deciding.
-            return .incomplete
-        }
-
-        switch bytes[start + 1] {
-        case rightBracket:
-            return oscColorReplySequence(in: bytes, at: start)
-        case leftBracket:
-            return csiProbeReplySequence(in: bytes, at: start)
-        default:
-            return .passThrough
-        }
-    }
-
-    private static func oscColorReplySequence(
-        in bytes: [UInt8],
-        at start: Int
-    ) -> SSHPTYAttachReconnectInputFilterSequenceMatch {
-        var cursor = start + 2
-        var command = [UInt8]()
-
-        while cursor < bytes.count {
-            let byte = bytes[cursor]
-            if byte == semicolon {
-                break
-            }
-            if byte < 0x30 || byte > 0x39 || command.count >= 2 {
-                return .passThrough
-            }
-            command.append(byte)
-            cursor += 1
-        }
-
-        guard cursor < bytes.count else {
-            return isOSCColorReplyCommandPrefix(command) ? .incomplete : .passThrough
-        }
-        guard bytes[cursor] == semicolon else {
-            return .passThrough
-        }
-        guard command == [0x31, 0x30] || command == [0x31, 0x31] || command == [0x31, 0x32] else {
-            return .passThrough
-        }
-
-        cursor += 1
-        while cursor < bytes.count {
-            let byte = bytes[cursor]
-            if byte == bell {
-                return .strip(length: cursor - start + 1)
-            }
-            if byte == escape {
-                guard cursor + 1 < bytes.count else {
-                    return .incomplete
-                }
-                if bytes[cursor + 1] == backslash {
-                    return .strip(length: cursor - start + 2)
-                }
-            }
-            cursor += 1
-        }
-        return .incomplete
-    }
-
-    private static func csiProbeReplySequence(
-        in bytes: [UInt8],
-        at start: Int
-    ) -> SSHPTYAttachReconnectInputFilterSequenceMatch {
-        var cursor = start + 2
-        while cursor < bytes.count {
-            let byte = bytes[cursor]
-            if byte >= 0x40, byte <= 0x7E {
-                return shouldStripCSIReply(bytes: bytes, bodyStart: start + 2, finalIndex: cursor)
-                    ? .strip(length: cursor - start + 1)
-                    : .passThrough
-            }
-            guard byte >= 0x20, byte <= 0x3F else {
-                return .passThrough
-            }
-            cursor += 1
-        }
-        return .incomplete
-    }
-
-    private static func isOSCColorReplyCommandPrefix(_ command: [UInt8]) -> Bool {
-        command.isEmpty ||
-            command == [0x31] ||
-            command == [0x31, 0x30] ||
-            command == [0x31, 0x31] ||
-            command == [0x31, 0x32]
-    }
-
-    private static func shouldStripCSIReply(bytes: [UInt8], bodyStart: Int, finalIndex: Int) -> Bool {
-        var parameterEnd = bodyStart
-        while parameterEnd < finalIndex, bytes[parameterEnd] >= 0x30, bytes[parameterEnd] <= 0x3F {
-            parameterEnd += 1
-        }
-        guard bytes[parameterEnd..<finalIndex].allSatisfy({ $0 >= 0x20 && $0 <= 0x2F }) else {
-            return false
-        }
-
-        let parameters = bytes[bodyStart..<parameterEnd]
-        let intermediates = bytes[parameterEnd..<finalIndex]
-        let final = bytes[finalIndex]
-
-        switch final {
-        case 0x49, 0x4F:
-            return parameters.isEmpty && intermediates.isEmpty
-        case 0x52, 0x63, 0x6E:
-            return intermediates.isEmpty
-        case 0x75:
-            return intermediates.isEmpty && parameters.first == questionMark
-        case 0x79:
-            return intermediates.elementsEqual([dollar])
-        default:
-            return false
-        }
-    }
-
 }
